Clarified map_coords
It's unclear to me that the map_coords function only results in applying a function to the individual dimensions of the tuple. This works well for basic scalling, but if you scale arbitrarily, you lose a lot of the 'geo' part of json, which is the geometric center issue.

I've also added a worked example of transforming coords in another pull request with the same scaffold of list comprehension.